[FEATURE-28] removed duplicated method
